// src/user/types/docente_aula.response.ts
// Generated by https://quicktype.io

export interface DocenteAulaResponse {
    id:      string;
    user:    User;
    aula:    Aula;
    curso:   Curso;
    materia: Materia;
    dia:     string;
    horario: string;
}

export interface Aula {
    id:          string;
    name:        string;
    description: string;
    capacity:    number;
    state:       string;
}

export interface Curso {
    id:               string;
    name:             string;
    shift:            string;
    numberOfStudents: number;
}

export interface Materia {
    id:   string;
    name: string;
}

export interface User {
    id:       string;
    name:     string;
    DNI:      number;
    roles:    string[];
    isActive: boolean;
}